created migration file

+++ app/models/reviews.py
@@ -14,7 +14,7 @@ class Review(db.Model, UserMixin):
     user_id = db.Column(db.Integer, db.ForeignKey(add_prefix_for_prod("users.id")), nullable=False)
     order_id = db.Column(db.Integer, db.ForeignKey(add_prefix_for_prod("orders.id")), nullable=False)
     review = db.Column(db.String(300), nullable=False)
-    star = db.Column(db.Integer(5), nullable=False)
+    star = db.Column(db.Integer(), nullable=False)
     created_at = db.Column(db.DateTime(timezone=True), default=db.func.now())
     updated_at = db.Column(db.DateTime(timezone=True), default=db.func.now())
